+++ Model/Plant/Phenology/Zadok.cpp
@@ -18,15 +18,15 @@ float Zadok::calcZadok()
    float fracInCurrent = plant.phenology().fractionInCurrentPhase();
    float zadok_stage = 0.0;
    if (plant.phenology().inPhase("sowing"))
-       zadok_stage = 5.0 * fracInCurrent;
+       zadok_stage = 5.0f * fracInCurrent;
    else if (plant.phenology().inPhase("germination"))
-       zadok_stage = 5.0 + 5 * fracInCurrent;
+       zadok_stage = 5.0f + 5 * fracInCurrent;
    else if (plant.phenology().inPhase("emergence"))
-       zadok_stage = 10.0;
+       zadok_stage = 10.0f;
    else if (plant.phenology().inPhase("end_of_juvenile") &&
             fracInCurrent <= 0.9)
       {
-      float leaf_no_now = max(0.0, plant.getLeafNo() - 2.0);
+      float leaf_no_now = (float)max(0.0, plant.getLeafNo() - 2.0f);
 
       static float tillerno_y[] =   // tiller no.
            {0, 0, 5};
@@ -37,9 +37,9 @@ float Zadok::calcZadok()
                                                , tillerno_y
                                                , sizeof(tillerno_x)/sizeof(float));
       if (tiller_no_now <= 0.0)
-         zadok_stage = 10.0 + leaf_no_now;
+         zadok_stage = 10.0f + leaf_no_now;
       else
-         zadok_stage = 20.0 + tiller_no_now;
+         zadok_stage = 20.0f + tiller_no_now;
       }
    else if (!plant.phenology().inPhase("out"))
       {
@@ -59,9 +59,9 @@ float Zadok::calcZadok()
       //                  na     na   na    na    na   30  43   65    71     87    90    100
 
       static float zadok_code_y[] =
-           {30,   40, 65, 71, 87, 90, 100};
+           {30.0f,   40.0f, 65.0f, 71.0f, 87.0f, 90.0f, 100.0f};
       static float zadok_code_x[] =
-           {4.9, 5.4,  6,  7,  8,  9,  10};
+           {4.9f, 5.4f,  6.0f,  7.0f,  8.0f,  9.0f,  10.0f};
       interpolationFunction zadok;
       zadok.setXY(vector<float> (zadok_code_x, zadok_code_x + sizeof(zadok_code_x)/sizeof(float)),
                   vector<float> (zadok_code_y, zadok_code_y + sizeof(zadok_code_y)/sizeof(float)));
